multimodal/Usecases/Facebook_AI_benchmark/requirements.txt
# Core
fedops
torch
torchvision
scikit-learn
pandas
numpy
tqdm

# NLP
transformers
datasets

# Image processing
Pillow
requests
